# packages/cactus-plugin-satp-hermes/src/main/json/openapi-api1-bundled.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SATP Gateway Client (AP1)",
    "version": "0.0.2",
    "description": "SATP is a protocol operating between two gateways that conducts the transfer of a digital asset from one gateway to another. The protocol establishes a secure channel between the endpoints and implements a 2-phase commit to ensure the properties of transfer atomicity, consistency, isolation and durability.\n\nThis API defines the gateway client facing API), API-Type 1 in the SATP-Core specification.\n\n**Additional Resources**:\n- [Proposed SATP Charter](https://datatracker.ietf.org/doc/charter-ietf-satp/)\n- [SATP Core draft](https://datatracker.ietf.org/doc/draft-ietf-satp-core)\n- [SATP Crash Recovery draft](https://datatracker.ietf.org/doc/draft-belchior-satp-gateway-recovery/)\n- [SATP Architecture draft](https://datatracker.ietf.org/doc/draft-ietf-satp-architecture/)\n- [SATP Use-Cases draft](https://datatracker.ietf.org/doc/draft-ramakrishna-sat-use-cases/)\n- [SATP Data sharing draft](https://datatracker.ietf.org/doc/draft-ramakrishna-satp-data-sharing)\n- [SATP View Addresses draft](https://datatracker.ietf.org/doc/draft-ramakrishna-satp-views-addresses)",
    "contact": {
      "name": "Hyperledger Cacti"
    },
    "x-logo": {
      "altText": "Cacti logo",
      "backgroundColor": "#fafafa",
      "url": "https://github.com/hyperledger/cacti/blob/main/images/HL_Cacti_Logo_Color.png"
    }
  },
...
  "servers": [
    {
...
      "variables": {
...
        },
...
  "paths": {
...
    "/api/v1/@hyperledger/cactus-plugin-satp-hermes/status": {
      "get": {
        "summary": "Get SATP current session data",
        "description": "Retrieve the status of a SATP session",
        "operationId": "GetStatus",
        "tags": [
          "admin"
        ],
        "x-hyperledger-cacti": {
          "http": {
            "verbLowerCase": "get",
            "path": "/api/v1/@hyperledger/cactus-plugin-satp-hermes/status"
          }
        },
        "parameters": [
          {
            "name": "SessionID",
            "in": "query",
            "schema": {
              "x-category": "satp-core",
              "type": "string",
              "description": "Unique identifier (UUID) for the session.",
              "example": "000003e8-e0b8-21ee-ba00-325096b39f47"
            },
            "required": true,
            "description": "Unique identifier for the session."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Provides the current status of the SATP session including detailed information on the progress, such as substatus, stage, and step, along with the session start time and chain information.",
                  "type": "object",
                  "x-category": "response",
                  "required": [
                    "status",
                    "substatus",
                    "stage",
                    "step",
                    "startTime",
                    "originNetwork",
                    "destinationNetwork"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "NOT_FOUND",
                        "INVALID",
                        "PENDING",
                        "DONE",
                        "FAILED"
                      ],
                      "example": "DONE"
                    },
                    "substatus": {
                      "type": "string",
                      "enum": [
                        "WAIT_SOURCE_CONFIRMATIONS",
                        "WAIT_DESTINATION_TRANSACTION",
                        "BRIDGE_NOT_AVAILABLE",
                        "CHAIN_NOT_AVAILABLE",
                        "REFUND_IN_PROGRESS",
                        "UNKNOWN_ERROR",
                        "COMPLETED",
                        "PARTIAL",
                        "REFUNDED",
                        "NOT_PROCESSABLE_REFUND_NEEDED",
                        "RECOVERING",
                        "RECOVERED",
                        "RECOVERY_FAILED",
                        "REJECTED"
                      ],
                      "example": "COMPLETED"
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "STAGE_0",
                        "STAGE_1",
                        "STAGE_2",
                        "STAGE_3"
                      ],
                      "example": "STAGE_3"
                    },
                    "step": {
                      "type": "string",
                      "enum": [
                        "transfer-initialization-claims",
                        "conveyance-of-network-capabilities-and-parameters",
                        "transfer-proposal-message",
                        "transfer-proposal-receipt-message",
                        "transfer-proposal-reject-and-conditional-reject-message",
                        "transfer-commence-message",
                        "commence-response-message",
                        "lock-assertion-message",
                        "lock-assertion-receipt-message",
                        "commit-preparation-message",
                        "commit-ready-message",
                        "commit-final-assertion-message",
                        "commit-final-acknowledgement-receipt-message",
                        "transfer-complete-message",
                        "rollback-transfer-initialization-claims",
                        "rollback-conveyance-of-network-capabilities-and-parameters",
                        "rollback-transfer-proposal-message",
                        "rollback-transfer-proposal-receipt-message",
                        "rollback-transfer-proposal-reject-and-conditional-reject-message",
                        "rollback-transfer-commence-message",
                        "rollback-commence-response-message",
                        "rollback-lock-assertion-message",
                        "rollback-lock-assertion-receipt-message",
                        "rollback-commit-preparation-message",
                        "rollback-commit-ready-message",
                        "rollback-commit-final-assertion-message",
                        "rollback-commit-final-acknowledgement-receipt-message",
                        "rollback-transfer-complete-message"
                      ],
                      "example": "transfer-complete-message"
                    },
                    "startTime": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-01-01T00:00:00Z"
                    },
                    "originNetwork": {
                      "type": "object",
                      "properties": {
                        "dltProtocol": {
                          "example": "HyperledgerFabric"
                        },
                        "dltSubnetworkID": {
                          "example": "subnetwork1"
                        }
                      }
                    },
                    "destinationNetwork": {
                      "type": "object",
                      "properties": {
                        "dltProtocol": {
                          "example": "HyperledgerBesu"
                        },
                        "dltSubnetworkID": {
                          "example": "subnetwork2"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}